Basic video device support & bug fixes
The firmware itself does support the video device, but will only make
use of it if it is present, unlike the console and storage devices,
which are both required.

The test bootloader does not make use of the video device as it very
experimental in the emulator.

By default the video device is disabed.