Firmware Storage API
This is some very useful API that allows a bootloader to request a
storage read/write from the firmware. The test bootloader uses this to
load an example kernel which just prints a simple message to the
console.

There were some minor bug fixes to the firmware and test
bootloader with this commit.